@@ -1,298 +1,301 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup Label="MultilingualAppToolkit">
-        <MultilingualAppToolkitVersion>4.0</MultilingualAppToolkitVersion>
-        <MultilingualFallbackLanguage>en</MultilingualFallbackLanguage>
-        <TranslationReport Condition="'$(Configuration)' == 'Release'">true</TranslationReport>
-        <SuppressPseudoWarning Condition="'$(Configuration)' == 'Debug'">true</SuppressPseudoWarning>
-    </PropertyGroup>
-    <PropertyGroup>
-        <TargetFramework>net5.0-windows</TargetFramework>
-        <RuntimeIdentifier>win-x64</RuntimeIdentifier>
-        <AssemblyName>Echoglossian</AssemblyName>
-        <Platforms>x64</Platforms>
-        <Authors>lokinmodar</Authors>
-        <NeutralLanguage>en</NeutralLanguage>
-        <LangVersion>9.0</LangVersion>
-        <VersionMajor>2</VersionMajor>
-        <VersionMinor>2</VersionMinor>
-        <VersionPatch Condition="'$(VersionPatch)' == ''">$([System.DateTime]::UtcNow.ToString("yyMM"))</VersionPatch>
-        <VersionRevision Condition="'$(VersionRevision)' == ''">$([System.DateTime]::UtcNow.TimeOfDay.TotalMinutes.ToString("0"))</VersionRevision>
-        <Version>$(VersionMajor).$(VersionMinor).$(VersionPatch).$(VersionRevision)</Version>
-        <AssemblyVersion>$(Version)</AssemblyVersion>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-        <ProduceReferenceAssembly>false</ProduceReferenceAssembly>
-        <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
-        <NoWarn>IDE0003</NoWarn>
-        <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-        <Company>lokinmodar</Company>
-        <Description>Final Fantasy XIV game text translator</Description>
-        <Copyright>Dante Souza e Souza - aka lokinmodar - 2021</Copyright>
-        <PackageProjectUrl>https://github.com/lokinmodar/Echoglossian/</PackageProjectUrl>
-        <DalamudLibPath>$(appdata)\XIVLauncher\addon\Hooks\dev\</DalamudLibPath>
-        <PackageLicenseExpression>Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International Public License</PackageLicenseExpression>
-        <PackageIcon>logo.png</PackageIcon>
-        <PackageIconUrl>https://github.com/lokinmodar/Echoglossian/blob/APIv4/images/logo.png?raw=true</PackageIconUrl>
-        <RepositoryUrl>https://github.com/lokinmodar/Echoglossian</RepositoryUrl>
-        <PackageTags>ffxiv, translation, dll, plugin, XIVLauncher</PackageTags>
-        <DocumentationFile>$(OutputPath)$(AssemblyName).xml</DocumentationFile>
-    </PropertyGroup>
-    <Import Project="$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\Microsoft.Multilingual.ResxResources.targets" Label="MultilingualAppToolkit" Condition="Exists('$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\v$(MultilingualAppToolkitVersion)\Microsoft.Multilingual.ResxResources.targets')"/>
-    <Target Name="MATPrerequisite" BeforeTargets="PrepareForBuild" Condition="!Exists('$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\Microsoft.Multilingual.ResxResources.targets')" Label="MultilingualAppToolkit">
-        <Warning Text="$(MSBuildProjectFile) is Multilingual build enabled, but the Multilingual App Toolkit is unavailable during the build. If building with Visual Studio, please check to ensure that toolkit is properly installed."/>
-    </Target>
-    <PropertyGroup>
-        <DalamudLibPath>$(appdata)\XIVLauncher\addon\Hooks\dev\</DalamudLibPath>
-    </PropertyGroup>
-
-
-    <ItemGroup>
-        <Reference Include="Dalamud">
-            <HintPath>$(DalamudLibPath)Dalamud.dll</HintPath>
-            <Private>false</Private>
-        </Reference>
-        <Reference Include="FFXIVClientStructs">
-            <HintPath>$(DalamudLibPath)FFXIVClientStructs.dll</HintPath>
-        </Reference>
-        <Reference Include="GTranslate">
-            <HintPath>I:\Dante\tataru-to-dalamud\GTranslate\src\GTranslate\obj\Debug\net5.0-windows7.0\GTranslate.dll</HintPath>
-            <Private>true</Private>
-        </Reference>
-        <Reference Include="ImGui.NET">
-            <HintPath>$(DalamudLibPath)ImGui.NET.dll</HintPath>
-            <Private>false</Private>
-        </Reference>
-        <Reference Include="ImGuiScene">
-            <HintPath>$(DalamudLibPath)ImGuiScene.dll</HintPath>
-            <Private>false</Private>
-        </Reference>
-        <Reference Include="JetBrains.Annotations, Version=2021.2.0.0, Culture=neutral, PublicKeyToken=1010a0d8d6380325">
-            <HintPath>C:\Users\lokin\AppData\Roaming\XIVLauncher\addon\Hooks\dev\JetBrains.Annotations.dll</HintPath>
-        </Reference>
-        <Reference Include="Lumina">
-            <HintPath>$(DalamudLibPath)Lumina.dll</HintPath>
-            <Private>false</Private>
-        </Reference>
-        <Reference Include="Lumina.Excel">
-            <HintPath>$(DalamudLibPath)Lumina.Excel.dll</HintPath>
-            <Private>false</Private>
-        </Reference>
-    </ItemGroup>
-    <ItemGroup>
-        <None Update="Font\symbols.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </None>
-    </ItemGroup>
-
-    <ItemGroup>
-        <AdditionalFiles Include=".\stylecop.settings"/>
-        <AdditionalFiles Include="stylecop.settings"/>
-    </ItemGroup>
-    <ItemGroup>
-        <None Remove="Font\NotoSans-Medium-Custom2.otf"/>
-        <None Remove="Font\NotoSans-Medium.ttf"/>
-        <None Remove="Font\NotoSans-Regular.ttf"/>
-        <None Remove="Font\NotoSans-SemiBold.ttf"/>
-        <None Remove="Font\NotoSansArabic-Medium.ttf"/>
-        <None Remove="Font\NotoSansArmenian-Medium.ttf"/>
-        <None Remove="Font\NotoSansBengali-Medium.ttf"/>
-        <None Remove="Font\NotoSansCJKhk-Regular.otf"/>
-        <None Remove="Font\NotoSansCJKjp-Regular.otf"/>
-        <None Remove="Font\NotoSansCJKkr-Regular.otf"/>
-        <None Remove="Font\NotoSansCJKsc-Regular.otf"/>
-        <None Remove="Font\NotoSansCJKtc-Regular.otf"/>
-        <None Remove="Font\NotoSansDevanagari-Medium.ttf"/>
-        <None Remove="Font\NotoSansGeorgian-Medium.ttf"/>
-        <None Remove="Font\NotoSansGlagolitic-Regular.ttf"/>
-        <None Remove="Font\NotoSansGujarati-Medium.ttf"/>
-        <None Remove="Font\NotoSansGurmukhi-Medium.ttf"/>
-        <None Remove="Font\NotoSansHebrew-Medium.ttf"/>
-        <None Remove="Font\NotoSansJavanese-Regular.ttf"/>
-        <None Remove="Font\NotoSansKannada-Medium.ttf"/>
-        <None Remove="Font\NotoSansKhmer-Medium.ttf"/>
-        <None Remove="Font\NotoSansLao-Medium.ttf"/>
-        <None Remove="Font\NotoSansMalayalam-Medium.ttf"/>
-        <None Remove="Font\NotoSansMongolian-Regular.ttf"/>
-        <None Remove="Font\NotoSansMyanmar-Medium.ttf"/>
-        <None Remove="Font\NotoSansOriya-Regular.ttf"/>
-        <None Remove="Font\NotoSansSinhala-Medium.ttf"/>
-        <None Remove="Font\NotoSansTagalog-Regular.ttf"/>
-        <None Remove="Font\NotoSansTamil-Medium.ttf"/>
-        <None Remove="Font\NotoSansTelugu-Medium.ttf"/>
-        <None Remove="Font\NotoSansThai-Medium.ttf"/>
-        <None Remove="images\icon.png"/>
-        <None Remove="images\logo.png"/>
-        <None Remove="minfilia.ico"/>
-
-
-        <None Remove="MultilingualResources\Echoglossian.da.xlf"/>
-        <None Remove="MultilingualResources\Echoglossian.de.xlf"/>
-        <None Remove="MultilingualResources\Echoglossian.el.xlf"/>
-        <None Remove="MultilingualResources\Echoglossian.es.xlf"/>
-        <None Remove="MultilingualResources\Echoglossian.fi.xlf"/>
-        <None Remove="MultilingualResources\Echoglossian.fil.xlf"/>
-        <None Remove="MultilingualResources\Echoglossian.fr.xlf"/>
-        <None Remove="MultilingualResources\Echoglossian.ga.xlf"/>
-        <None Remove="MultilingualResources\Echoglossian.pt-BR.xlf"/>
-        <None Remove="MultilingualResources\Echoglossian.ru.xlf"/>
-
-
-        <None Remove="Resources\pix.png"/>
-        <None Remove="stylecop.json"/>
-        <None Remove="Wiki82.profile.xml"/>
-    </ItemGroup>
-    <ItemGroup>
-        <Content Include="Font\NotoSans-Medium-Custom2.otf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSans-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSans-SemiBold.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansArabic-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansArmenian-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansBengali-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansCJKhk-Regular.otf">
-            <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansCJKjp-Regular.otf">
-            <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansCJKkr-Regular.otf">
-            <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansCJKsc-Regular.otf">
-            <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansCJKtc-Regular.otf">
-            <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansDevanagari-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansGeorgian-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansGlagolitic-Regular.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansGujarati-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansGurmukhi-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansHebrew-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansJavanese-Regular.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansKannada-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansKhmer-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansLao-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansMalayalam-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansMongolian-Regular.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansMyanmar-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansOriya-Regular.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansSinhala-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansTagalog-Regular.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansTamil-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansTelugu-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Font\NotoSansThai-Medium.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="images\icon.png">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="Wiki82.profile.xml">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-    </ItemGroup>
-    <ItemGroup>
-        <PackageReference Include="AWSSDK.Translate" Version="3.7.3.16"/>
-        <PackageReference Include="DalamudLinter" Version="1.0.3"/>
-        <PackageReference Include="DalamudPackager" Version="2.1.5"/>
-        <PackageReference Include="DeepL" Version="0.4.0"/>
-
-        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.11"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="5.0.11"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="5.0.11">
-            <PrivateAssets>all</PrivateAssets>
-            <!--<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>-->
-        </PackageReference>
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.1"/>
-        <PackageReference Include="NTextCat" Version="0.3.65"/>
-        <PackageReference Include="Siggingway" Version="1.1.1"/>
-        <PackageReference Include="System.Drawing.Common" Version="6.0.0"/>
-        <PackageReference Include="System.Drawing.Primitives" Version="4.3.0"/>
-        <PackageReference Include="XivCommon" Version="5.0.0"/>
-    </ItemGroup>
-    <ItemGroup>
-        <Content Include="Font\NotoSans-Regular.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-    </ItemGroup>
-    <ItemGroup>
-        <EmbeddedResource Include="images\logo.png">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </EmbeddedResource>
-        <EmbeddedResource Include="Resources\pix.png"/>
-        <EmbeddedResource Include="Resources\minfilia.ico"/>
-        <XliffResource Include="MultilingualResources\Echoglossian.da.xlf"/>
-        <XliffResource Include="MultilingualResources\Echoglossian.de.xlf"/>
-        <XliffResource Include="MultilingualResources\Echoglossian.es.xlf"/>
-        <XliffResource Include="MultilingualResources\Echoglossian.fr.xlf"/>
-        <XliffResource Include="MultilingualResources\Echoglossian.pt-BR.xlf"/>
-    </ItemGroup>
-    <ItemGroup>
-        <Compile Update="Properties\Resources.Designer.cs">
-            <DesignTime>True</DesignTime>
-            <AutoGen>True</AutoGen>
-            <DependentUpon>Resources.resx</DependentUpon>
-        </Compile>
-        <Compile Remove="Class1.cs"/>
-    </ItemGroup>
-    <ItemGroup>
-        <EmbeddedResource Update="Properties\Resources.resx">
-            <Generator>PublicResXFileCodeGenerator</Generator>
-            <LastGenOutput>Resources.Designer.cs</LastGenOutput>
-        </EmbeddedResource>
-    </ItemGroup>
-    <ItemGroup>
-        <None Update="Font\symbols.ttf">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </None>
-    </ItemGroup>
+  <PropertyGroup Label="MultilingualAppToolkit">
+    <MultilingualAppToolkitVersion>4.0</MultilingualAppToolkitVersion>
+    <MultilingualFallbackLanguage>en</MultilingualFallbackLanguage>
+    <TranslationReport Condition="'$(Configuration)' == 'Release'">true</TranslationReport>
+    <SuppressPseudoWarning Condition="'$(Configuration)' == 'Debug'">true</SuppressPseudoWarning>
+  </PropertyGroup>
+  <PropertyGroup>
+    <TargetFramework>net5.0-windows7.0</TargetFramework>
+    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
+    <AssemblyName>Echoglossian</AssemblyName>
+    <Platforms>x64</Platforms>
+    <Authors>lokinmodar</Authors>
+    <NeutralLanguage>en</NeutralLanguage>
+    <LangVersion>10.0</LangVersion>
+    <VersionMajor>3</VersionMajor>
+    <VersionMinor>0</VersionMinor>
+    <VersionPatch Condition="'$(VersionPatch)' == ''">$([System.DateTime]::UtcNow.ToString("yyMM"))</VersionPatch>
+    <VersionRevision Condition="'$(VersionRevision)' == ''">$([System.DateTime]::UtcNow.TimeOfDay.TotalMinutes.ToString("0"))</VersionRevision>
+    <Version>$(VersionMajor).$(VersionMinor).$(VersionPatch).$(VersionRevision)</Version>
+    <AssemblyVersion>$(Version)</AssemblyVersion>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <ProduceReferenceAssembly>false</ProduceReferenceAssembly>
+    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
+    <NoWarn>IDE0003</NoWarn>
+    <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
+    <Company>lokinmodar</Company>
+    <Description>Final Fantasy XIV game text translator</Description>
+    <Copyright>Dante Souza e Souza - aka lokinmodar - 2022</Copyright>
+    <PackageProjectUrl>https://github.com/lokinmodar/Echoglossian/</PackageProjectUrl>
+    <DalamudLibPath>$(appdata)\XIVLauncher\addon\Hooks\dev\</DalamudLibPath>
+    <ApplicationIcon>minfilia.ico</ApplicationIcon>
+    <PackageLicenseExpression>Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International Public License</PackageLicenseExpression>
+    <PackageIcon>logo.png</PackageIcon>
+    <PackageIconUrl>https://github.com/lokinmodar/Echoglossian/blob/APIv4/images/logo.png?raw=true</PackageIconUrl>
+    <RepositoryUrl>https://github.com/lokinmodar/Echoglossian</RepositoryUrl>
+    <PackageTags>ffxiv, translation, dll, plugin, XIVLauncher</PackageTags>
+    <DocumentationFile>$(OutputPath)$(AssemblyName).xml</DocumentationFile>
+  </PropertyGroup>
+  <Import Project="$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\Microsoft.Multilingual.ResxResources.targets" Label="MultilingualAppToolkit" Condition="Exists('$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\v$(MultilingualAppToolkitVersion)\Microsoft.Multilingual.ResxResources.targets')" />
+  <Target Name="MATPrerequisite" BeforeTargets="PrepareForBuild" Condition="!Exists('$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\Microsoft.Multilingual.ResxResources.targets')" Label="MultilingualAppToolkit">
+    <Warning Text="$(MSBuildProjectFile) is Multilingual build enabled, but the Multilingual App Toolkit is unavailable during the build. If building with Visual Studio, please check to ensure that toolkit is properly installed." />
+  </Target>
+  <PropertyGroup>
+    <DalamudLibPath>$(appdata)\XIVLauncher\addon\Hooks\dev\</DalamudLibPath>
+  </PropertyGroup>
+  <ItemGroup>
+    <AdditionalFiles Include=".\stylecop.json" />
+    <AdditionalFiles Include="stylecop.json" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="Font\NotoSans-Medium-Custom2.otf" />
+    <None Remove="Font\NotoSans-Medium.ttf" />
+    <None Remove="Font\NotoSans-Regular.ttf" />
+    <None Remove="Font\NotoSans-SemiBold.ttf" />
+    <None Remove="Font\NotoSansArabic-Medium.ttf" />
+    <None Remove="Font\NotoSansArmenian-Medium.ttf" />
+    <None Remove="Font\NotoSansBengali-Medium.ttf" />
+    <None Remove="Font\NotoSansCJKhk-Regular.otf" />
+    <None Remove="Font\NotoSansCJKjp-Regular.otf" />
+    <None Remove="Font\NotoSansCJKkr-Regular.otf" />
+    <None Remove="Font\NotoSansCJKsc-Regular.otf" />
+    <None Remove="Font\NotoSansCJKtc-Regular.otf" />
+    <None Remove="Font\NotoSansDevanagari-Medium.ttf" />
+    <None Remove="Font\NotoSansGeorgian-Medium.ttf" />
+    <None Remove="Font\NotoSansGlagolitic-Regular.ttf" />
+    <None Remove="Font\NotoSansGujarati-Medium.ttf" />
+    <None Remove="Font\NotoSansGurmukhi-Medium.ttf" />
+    <None Remove="Font\NotoSansHebrew-Medium.ttf" />
+    <None Remove="Font\NotoSansJavanese-Regular.ttf" />
+    <None Remove="Font\NotoSansKannada-Medium.ttf" />
+    <None Remove="Font\NotoSansKhmer-Medium.ttf" />
+    <None Remove="Font\NotoSansLao-Medium.ttf" />
+    <None Remove="Font\NotoSansMalayalam-Medium.ttf" />
+    <None Remove="Font\NotoSansMongolian-Regular.ttf" />
+    <None Remove="Font\NotoSansMyanmar-Medium.ttf" />
+    <None Remove="Font\NotoSansOriya-Regular.ttf" />
+    <None Remove="Font\NotoSansSinhala-Medium.ttf" />
+    <None Remove="Font\NotoSansTagalog-Regular.ttf" />
+    <None Remove="Font\NotoSansTamil-Medium.ttf" />
+    <None Remove="Font\NotoSansTelugu-Medium.ttf" />
+    <None Remove="Font\NotoSansThai-Medium.ttf" />
+    <None Remove="images\icon.png" />
+    <None Remove="images\logo.png" />
+    <None Remove="minfilia.ico" />
+    <None Remove="MultilingualResources\Echoglossian.ca.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.cs.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.da.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.de.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.el.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.es.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.fi.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.fil.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.fr.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.ga.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.gl.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.gu.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.hi.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.hu.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.ja.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.ko.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.ms.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.pa-Guru.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.pl.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.pt-BR.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.ru.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.sd-Deva.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.ta.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.vi.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.zh-Hans.xlf" />
+    <None Remove="MultilingualResources\Echoglossian.zh-Hant.xlf" />
+    <None Remove="PluginUI - Copy___" />
+    <None Remove="Resources\pix.png" />
+    <None Remove="stylecop.json" />
+    <None Remove="Wiki82.profile.xml" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Font\NotoSans-Medium-Custom2.otf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSans-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSans-SemiBold.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansArabic-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansArmenian-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansBengali-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansCJKhk-Regular.otf">
+      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansCJKjp-Regular.otf">
+      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansCJKkr-Regular.otf">
+      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansCJKsc-Regular.otf">
+      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansCJKtc-Regular.otf">
+      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansDevanagari-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansGeorgian-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansGlagolitic-Regular.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansGujarati-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansGurmukhi-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansHebrew-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansJavanese-Regular.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansKannada-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansKhmer-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansLao-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansMalayalam-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansMongolian-Regular.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansMyanmar-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansOriya-Regular.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansSinhala-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansTagalog-Regular.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansTamil-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansTelugu-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Font\NotoSansThai-Medium.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="images\icon.png">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Wiki82.profile.xml">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="AWSSDK.Translate" Version="3.7.4.4" />
+    <PackageReference Include="DalamudLinter" Version="1.0.3" />
+    <PackageReference Include="DalamudPackager" Version="2.1.7" />
+    <PackageReference Include="DeepL" Version="0.4.0" />
+    <PackageReference Include="GTranslate" Version="2.0.2" />
+    <PackageReference Include="HtmlAgilityPack" Version="1.11.42" />
+    <PackageReference Include="Lumina.Excel" Version="6.1.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.11" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="5.0.11" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="5.0.11">
+      <PrivateAssets>all</PrivateAssets>
+      <!--<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>-->
+    </PackageReference>
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="NTextCat" Version="0.3.65" />
+    <PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
+      <PrivateAssets>all</PrivateAssets>
+      <!--<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>-->
+    </PackageReference>
+    <PackageReference Include="System.Drawing.Common" Version="6.0.0-rc.2.21480.5" />
+    <PackageReference Include="System.Drawing.Primitives" Version="4.3.0" />
+    <PackageReference Include="XivCommon" Version="5.0.0" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Font\NotoSans-Regular.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="images\logo.png">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Resources\pix.png" />
+    <EmbeddedResource Include="Resources\minfilia.ico" />
+    <XliffResource Include="MultilingualResources\Echoglossian.da.xlf" />
+    <XliffResource Include="MultilingualResources\Echoglossian.de.xlf" />
+    <XliffResource Include="MultilingualResources\Echoglossian.es.xlf" />
+    <XliffResource Include="MultilingualResources\Echoglossian.fr.xlf" />
+    <XliffResource Include="MultilingualResources\Echoglossian.pt-BR.xlf" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Update="Properties\Resources.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>Resources.resx</DependentUpon>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Update="Properties\Resources.resx">
+      <Generator>PublicResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <Reference Include="Dalamud">
+      <HintPath>$(DalamudLibPath)Dalamud.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="FFXIVClientStructs">
+      <HintPath>$(DalamudLibPath)FFXIVClientStructs.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="ImGui.NET">
+      <HintPath>$(DalamudLibPath)ImGui.NET.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="ImGuiScene">
+      <HintPath>$(DalamudLibPath)ImGuiScene.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="Lumina">
+      <HintPath>$(DalamudLibPath)Lumina.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="Lumina.Excel">
+      <HintPath>$(DalamudLibPath)Lumina.Excel.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+  </ItemGroup>
+  <ItemGroup>
+    <None Update="Font\symbols.ttf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
 
 </Project>
